chore: move image package to client
Move image package to client because it's now used both omni and omni/client
